backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
pdfplumber==0.11.0
python-docx==1.1.0
reportlab==4.2.0
anthropic>=0.50.0
openai==1.50.0
httpx>=0.25.0,<0.28.0
python-dotenv==1.0.0
pydantic==2.11.3
slowapi==0.1.9
json-repair>=0.30.0
